test: Add test for service migration between legacy and v2
The test is based on a dummy TCP listener and sender [1] which
panic if recv or send fails. Running those in containers make
a container to restart if there were any failures. So, by
inspecting restart count we can see whether there were any
interruptions in established connections after we migrate
between legacy and v2 services.

We set bpf-ct-global-tcp-max=1000000 to make sure that CT does not
get recreated due to different size to avoid losing connections.

[1]: https://github.com/cilium/migrate-svc-test

Signed-off-by: Martynas Pumputis <m@lambda.lt>